+++ src/main/scala/com/caju/authorizer/repository/AccountRepository.scala
@@ -4,7 +4,7 @@ import zio.*
 import zio.schema.*
 import zio.schema.DeriveSchema.*
 
-case class Account(id: String, balanceFood: Long, balanceMeal: Long, balanceCash: Long)
+case class Account(id: String, balanceFood: Double, balanceMeal: Double, balanceCash: Double)
 
 object Account:
 	given Schema[Account] = DeriveSchema.gen[Account]
@@ -14,11 +14,11 @@ trait AccountRepository:
 
   def lookup(id: String): Task[Option[Account]]
 
-  def updateFoodBalance(user: Account, balance: Long): Task[String]
+  def updateFoodBalance(user: Account, balance: Double): Task[String]
 
-  def updateMealBalance(user: Account, balance: Long): Task[String]
+  def updateMealBalance(user: Account, balance: Double): Task[String]
 
-  def updateCashBalance(user: Account, balance: Long): Task[String]
+  def updateCashBalance(user: Account, balance: Double): Task[String]
 
 object AccountRepository:
   def register(account: Account): ZIO[AccountRepository, Throwable, String] =
@@ -27,11 +27,11 @@ object AccountRepository:
   def lookup(id: String): ZIO[AccountRepository, Throwable, Option[Account]] =
     ZIO.serviceWithZIO[AccountRepository](_.lookup(id))
 
-  def updateFoodBalance(account: Account, balance: Long): ZIO[AccountRepository, Throwable, String] =
+  def updateFoodBalance(account: Account, balance: Double): ZIO[AccountRepository, Throwable, String] =
     ZIO.serviceWithZIO[AccountRepository](_.updateFoodBalance(account, balance))
 
-  def updateMealBalance(account: Account, balance: Long): ZIO[AccountRepository, Throwable, String] =
+  def updateMealBalance(account: Account, balance: Double): ZIO[AccountRepository, Throwable, String] =
     ZIO.serviceWithZIO[AccountRepository](_.updateMealBalance(account, balance))
 
-  def updateCashBalance(account: Account, balance: Long): ZIO[AccountRepository, Throwable, String] =
+  def updateCashBalance(account: Account, balance: Double): ZIO[AccountRepository, Throwable, String] =
     ZIO.serviceWithZIO[AccountRepository](_.updateCashBalance(account, balance))
